requirements.txt
fastapi===0.108.0
uvicorn===0.25.0
sqlalchemy===2.0.25
passlib===1.7.4
bcrypt===4.1.2
python-multipart===0.0.6
python-jose===3.3.0